% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{citrulline_data}
\alias{citrulline_data}
\title{citrullline_data; Data frame containing the data for the paper}
\format{A tibble}
\usage{
citrulline_data
}
\description{
Object is a tibble (tbl_df)
}
\examples{
head(citrulliner::citrulline_data)
library(citrulliner); data("citrulline_data")

}
\keyword{datasets}
